requirements.txt
psycopg2-binary
pandas
SQLAlchemy
scikit-learn
scipy
tensorboardX
joblib
